Enable peeking ILGen for methods with resolved invokedynamic/handle
This would allow object info propagation in cases where, for example,
the necessary object info is obtained from a field prior to the
invokedynamic/invokehandle operation.

There is a temporary workaround in place due to a functional issue
we may run into during the compilation of
javax/imageio/stream/ImageInputStreamImpl.readLong()J. This functional
issue surfaced as a result of the additional inlining of VarHandle
operations that is made possible through the peeking ILGen. This
workaround will be removed once the underlying issue is resolved.

This commit removes the need to enable this peeking behaviour using
env variable TR_enablePeekingForMHInvokes. With the workaround in
place, the peeking ILGen for invokedynamic/invokehandle-containing
methods are now enabled by default.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>