Free monitorEnterRecordPool in freecontinuation
Currently, the monitor enter record pool is only free'd when the
continuation is removed from the continuation cache. This behaviour
causes memory leaks as a continuation may never be removed from the
cache.

Additionally, when allocate a new continuation each time
enterContinuation is called.

This PR takes the conservative approach of free the monitor records as
soon as the continuation is free'd. In future we should consider if the
monitor records can be persisted run to run. Currently, this has not
been tested yet.

Fixes: https://github.com/eclipse-openj9/openj9/issues/22776

Signed-off-by: tajila <atobia@ca.ibm.com>